A4/requirements.txt
ufal.udpipe==1.3.0.1